requirements.txt
python-dotenv~=1.0.1
aiogram~=3.17.0